feat: prevent user from adding a past date
